% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slxr-package.R
\docType{package}
\name{slxr-package}
\alias{slxr}
\alias{slxr-package}
\title{slxr: Spatial-X (SLX) Models for Applied Researchers}
\description{
Tools for estimating, interpreting, and visualizing Spatial-X regression
models, with first-class support for variable-specific weights matrices,
higher-order spatial lags, and temporally-lagged spatial variables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cwimpy/slxr}
  \item Report bugs at \url{https://github.com/cwimpy/slxr/issues}
}

}
\author{
\strong{Maintainer}: Cameron Wimpy \email{cwimpy@astate.edu}

}
\keyword{internal}
